# specs/main/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Group Scoring System - Authentication API
  description: API for user registration, login, and authentication
  version: 1.0.0
  contact:
    name: Group Scoring System
    url: http://localhost:3000

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://yourdomain.com
    description: Production server

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email/password and default USER role
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
            example:
              name: "John Doe"
              email: "john@example.com"
              password: "SecurePass123!"
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
              example:
                success: true
                message: "User registered successfully"
                user:
                  id: "clp123abc"
                  name: "John Doe"
                  email: "john@example.com"
                  role: "USER"
                  createdAt: "2025-10-31T09:28:34.000Z"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "VALIDATION_ERROR"
                message: "Invalid email format"
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "USER_EXISTS"
                message: "Email already registered"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/reset-password:
    post:
      summary: Request password reset
      description: Send password reset email to user
      operationId: resetPassword
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
            example:
              email: "john@example.com"
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
              example:
                success: true
                message: "Password reset email sent"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    RegistrationRequest:
      type: object
      required:
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: "john@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 100
          description: User's password (minimum 8 chars, must include uppercase, lowercase, number, special char)
          example: "SecurePass123!"
      additionalProperties: false

    RegistrationResponse:
      type: object
      required:
        - success
        - user
      properties:
        success:
          type: boolean
          description: Registration success status
          example: true
        message:
          type: string
          description: Success message
          example: "User registered successfully"
        user:
          $ref: '#/components/schemas/User'

    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"

    PasswordResetResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Password reset email sent"

    User:
      type: object
      required:
        - id
        - email
        - role
        - createdAt
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "clp123abc"
        name:
          type: string
          nullable: true
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        role:
          type: string
          enum: ["USER", "ADMIN"]
          description: User's role
          example: "USER"
        createdAt:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2025-10-31T09:28:34.000Z"

    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          enum:
            - VALIDATION_ERROR
            - USER_EXISTS
            - INVALID_CREDENTIALS
            - SERVER_ERROR
            - NOT_FOUND
          description: Error type code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid email format"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests

security:
  - bearerAuth: []